pappy-188956aa9fa00b6b7b7822a90c068c241df0a765/requirements.txt
Django==5.0.2
Pillow==11.1.0
python-dotenv==1.0.1
django-crispy-forms==2.3
crispy-bootstrap5==2024.10
django-cleanup==9.0.0
django-debug-toolbar==5.0.1
django-otp==1.5.4
qrcode==8.0
phonenumbers==8.13.53
twilio==9.4.3
unidecode==1.3.8 